# Properly copy the replicas from the range descriptor.