--- a/kv/dist_sender.go
+++ b/kv/dist_sender.go
@@ -386,8 +386,7 @@ func (ds *DistSender) sendRPC(desc *proto.RangeDescriptor,
 	}
 
 	// Copy and rearrange the replicas suitably, then return the desired order.
-	replicas := proto.ReplicaSlice(append(make([]proto.Replica,
-		len(desc.Replicas)), desc.Replicas...))
+	replicas := proto.ReplicaSlice(append([]proto.Replica(nil), desc.Replicas...))
 	// Rearrange the replicas so that those replicas with long common
 	// prefix of attributes end up first. If there's no prefix, this is a
 	// no-op.