The previous code was prefixing the copied range descriptors with
num-replicas "zero" replicas. The lookup of the replica address in the
gossip info for NodeID 0 was actually showing up in benchmarks. The
below numbers were gathered using RPCSender, but a similar speedup
occurs for the HTTPSender.

benchmark                           old ns/op     new ns/op     delta
BenchmarkClientScan1Version1Row     359020        340027        -5.29%

benchmark                           old MB/s     new MB/s     speedup
BenchmarkClientScan1Version1Row     2.85         3.01         1.06x